refactor(fast_io): split sendfile.rs into per-concern submodules
sendfile.rs hit 1037 lines, well above the 650-line production cap.
Split it via the SPL pattern into a `sendfile/` directory:

- mod.rs: module docs, constants, public dispatch (send_file_to_writer,
  send_file_to_fd, send_file_to_fd_with_policy) and re-exports.
- linux.rs: Linux `try_sendfile` zero-copy implementation.
- macos.rs: Darwin `try_sendfile_macos` (with explicit lseek bookkeeping).
- fallback.rs: portable buffered read/write loops (`copy_via_readwrite`
  and unix `copy_via_fd_write`, deduplicated from the previous Linux /
  non-Linux copies).
- tests.rs: cross-platform writer tests plus the gated Linux and macOS
  socketpair/pipe tests.

Public API at `fast_io::sendfile::*` is unchanged, all existing
`#[allow(unsafe_code)]` attributes are preserved exactly, and every
submodule sits comfortably under the 650-line ceiling.
